Add contextual menu for TransactionsTableView
