+++ MoneyM/Scenes/Home/HomeViewController.swift
@@ -36,7 +36,7 @@ class HomeViewController: UIViewController {
 
 	var router: HomeRoutingLogic?
 
-	private var transactionsArray: [Home.TransactionTableViewCellModel] = []
+	private(set) var transactionsArray: [Home.TransactionTableViewCellModel] = []
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -165,74 +165,6 @@ extension HomeViewController: HomeDisplayLogic {
 		datePickerButton.setTitle(viewModel.title, for: .normal)
 	}
 
-}
-
-// MARK: - Table View Delegate
-extension HomeViewController: UITableViewDelegate, UITableViewDataSource {
-
-	func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-		scrollViewHeightConstraint.constant = tableView.contentSize.height
-	}
-
-	func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-		transactionsArray[section].transactions.count
-	}
-
-	func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-		let transaction = transactionsArray[indexPath.section].transactions[indexPath.row]
-		let cell: UITableViewCell?
-
-		if transaction.note.isEmpty {
-			cell = tableView.dequeueReusableCell(withIdentifier: "defaultCell")
-		} else {
-			cell = tableView.dequeueReusableCell(withIdentifier: "cellWithNote")
-		}
-
-		if let defaultCell = cell as? TransactionTableViewCell {
-			defaultCell.loadTransaction(transaction: transaction)
-		} else if let cellWithNote = cell as? TransactionTableViewCellNote {
-			cellWithNote.loadTransaction(transaction: transaction)
-		}
-
-		return cell ?? UITableViewCell()
-	}
-
-	func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-
-		let deleteAction = UIContextualAction(style: .destructive,
-											  title: NSLocalizedString("delete.title", comment: "")) { _, _, _ in
-			let transaction = self.transactionsArray[indexPath.section].transactions[indexPath.row]
-			let request = Home.RemoveTransaction.Request(transaction: transaction)
-			self.interactor?.removeTransaction(request: request)
-		}
-
-		let editAction = UIContextualAction(style: .normal,
-											title: NSLocalizedString("edit.title", comment: "")) { _, _, _ in
-			let transaction = self.transactionsArray[indexPath.section].transactions[indexPath.row]
-			self.router?.routeToEditTransaction(transaction: transaction)
-		}
-		editAction.backgroundColor = .systemBlue
-
-		let swipeAction = UISwipeActionsConfiguration(actions: [deleteAction, editAction])
-
-		return swipeAction
-	}
-
-	func numberOfSections(in tableView: UITableView) -> Int {
-		transactionsArray.count
-	}
-
-	func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-		return transactionsArray[section].section
-	}
-
-	func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-		20
-	}
-
-	func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-		65
-	}
 }
 
 // MARK: - Add transaction delegate
